use an onboarding parent component
